requirements.txt
selenium
requests
mysql-connector
pyvirtualdisplay